requirements.txt
llama-index
llama-index-vector-stores-qdrant
llama-index-multi-modal-llms-openai
llama-index-readers-file
qdrant-client
matplotlib
pillow
pymupdf
torchvision
torch
transformers
python-dotenv
pandas
